+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   **Bolt** Added namespace config and secrets sync with `bolt config pull` and `bolt config push` via 1Password
 -   `GROUP_DEACTIVATED` error now shows reasons for deactivation. Added docs for deactivation reasons
 -   `/health/essential` endpoint to test connectivity to all essential services
+-   Added error when trying to deploy a distributed cluster on a non-linux-x86 machine (not supported)
 
 ### Changed